# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "AlertPage": {
        "additionalProperties": false,
        "description": "``GET /api/alerts`` — the only paging envelope in the API.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Alert"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor",
          "limit"
        ],
        "title": "AlertPage",
        "type": "object"
      },
...
      "ConfigPatch": {
        "additionalProperties": false,
        "description": "``PUT /api/config`` body. ``replay.speed`` is not patchable (R12).",
        "properties": {
          "values": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "integer"
                },
                {
                  "items": {
                    "type": "number"
                  },
                  "type": "array"
                }
              ]
            },
            "title": "Values",
            "type": "object"
          }
        },
        "required": [
          "values"
        ],
        "title": "ConfigPatch",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "REST surface of the explainable predictive-maintenance backend.\n\nTwo clocks are present on every payload: `ts` is wall-clock UTC and `dataset_ts`\nis the simulated instant the row represents. Both are ISO-8601 strings; every\nquery, window, percentile, streak and scrub uses `dataset_ts`.\n\nAll SHAP quantities are in probability space, so a waterfall sums from\n`base_value` to the probability shown on screen. Errors are RFC-9457\n`application/problem+json` bodies.",
    "title": "XPM — Explainable Predictive Maintenance API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/alerts": {
      "get": {
        "description": "The alert feed. The only paged endpoint in the API.",
        "operationId": "list_alerts_api_alerts_get",
        "parameters": [
          {
            "in": "query",
            "name": "plant_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "ai4i",
                    "ims"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Plant Id"
            }
          },
          {
            "in": "query",
            "name": "machine_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(ai4i|ims)-\\d{2}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Machine Id"
            }
          },
          {
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          },
          {
            "in": "query",
            "name": "until",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Until"
            }
          },
          {
            "in": "query",
            "name": "severity",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "medium",
                    "high",
                    "critical"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          },
          {
            "in": "query",
            "name": "feature",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Feature"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "summary": "List Alerts",
        "tags": [
          "alerts"
        ]
      }
    },
    "/api/alerts/{alert_id}": {
      "get": {
        "description": "One alert, open or closed.",
        "operationId": "get_alert_api_alerts__alert_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "alert_id",
            "required": true,
            "schema": {
              "pattern": "^alt_[0-9a-f]{16}$",
              "title": "Alert Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Alert"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "No such resource"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "summary": "Get Alert",
        "tags": [
          "alerts"
        ]
      }
    },
    "/api/alerts/{alert_id}/compare": {
      "get": {
        "description": "LightGBM against RandomForest for the same alert.",
        "operationId": "compare_models_api_alerts__alert_id__compare_get",
        "parameters": [
          {
            "in": "path",
            "name": "alert_id",
            "required": true,
            "schema": {
              "pattern": "^alt_[0-9a-f]{16}$",
              "title": "Alert Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelComparison"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "No such resource"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "summary": "Compare Models",
        "tags": [
          "explanations"
        ]
      }
    },
    "/api/alerts/{alert_id}/explanation": {
      "get": {
        "description": "The stored explanation for an alert. Never recomputed.",
        "operationId": "get_explanation_api_alerts__alert_id__explanation_get",
        "parameters": [
          {
            "in": "path",
            "name": "alert_id",
            "required": true,
            "schema": {
              "pattern": "^alt_[0-9a-f]{16}$",
              "title": "Alert Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "model",
            "required": false,
            "schema": {
              "default": "lgbm",
              "enum": [
                "lgbm",
                "rf"
              ],
              "title": "Model",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Explanation"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "No such resource"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "summary": "Get Explanation",
        "tags": [
          "explanations"
        ]
      }
    },
    "/api/config": {
      "get": {
        "description": "The full flattened settings tree plus the mutable-key list.",
        "operationId": "get_config_api_config_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "summary": "Get Config",
        "tags": [
          "config"
        ]
      },
      "put": {
        "description": "Patch mutable settings; mints a new run id and broadcasts a config frame.",
        "operationId": "put_config_api_config_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Key is not in mutable_keys"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "summary": "Put Config",
        "tags": [
          "config"
        ]
      }
    },
...
    "/api/machines": {
      "get": {
        "description": "Tile payloads for a plant. Bare list, no envelope.",
        "operationId": "list_machines_api_machines_get",
        "parameters": [
          {
            "in": "query",
            "name": "plant_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "ai4i",
                    "ims"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Plant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MachineSummary"
                  },
                  "title": "Response List Machines Api Machines Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "summary": "List Machines",
        "tags": [
          "machines"
        ]
      }
    },
    "/api/machines/{machine_id}": {
      "get": {
        "description": "One machine, including its channel list and alert count.",
        "operationId": "get_machine_api_machines__machine_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "machine_id",
            "required": true,
            "schema": {
              "pattern": "^(ai4i|ims)-\\d{2}$",
              "title": "Machine Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "No such resource"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "summary": "Get Machine",
        "tags": [
          "machines"
        ]
      }
    },
    "/api/machines/{machine_id}/importance": {
      "get": {
        "description": "Beeswarm-ready mean absolute SHAP across this machine's alerts.",
        "operationId": "get_importance_api_machines__machine_id__importance_get",
        "parameters": [
          {
            "in": "path",
            "name": "machine_id",
            "required": true,
            "schema": {
              "pattern": "^(ai4i|ims)-\\d{2}$",
              "title": "Machine Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          },
          {
            "in": "query",
            "name": "until",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Until"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalImportance"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "No such resource"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "summary": "Get Importance",
        "tags": [
          "explanations"
        ]
      }
    },
    "/api/models": {
      "get": {
        "description": "Registry metadata for both families. Bare list, no envelope.",
        "operationId": "list_models_api_models_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ModelInfo"
                  },
                  "title": "Response List Models Api Models Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "summary": "List Models",
        "tags": [
          "models"
        ]
      }
    },
    "/api/plants": {
      "get": {
        "description": "Every plant, with its canonical channel order. Bare list, no envelope.",
        "operationId": "list_plants_api_plants_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Plant"
                  },
                  "title": "Response List Plants Api Plants Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "summary": "List Plants",
        "tags": [
          "plants"
        ]
      }
    },
    "/api/replay": {
      "get": {
        "description": "Current transport state.",
        "operationId": "get_replay_api_replay_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplayState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "summary": "Get Replay",
        "tags": [
          "replay"
        ]
      }
    },
    "/api/replay/command": {
      "post": {
        "description": "The only authoritative transport control (R11).",
        "operationId": "post_replay_command_api_replay_command_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplayCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplayState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "summary": "Post Replay Command",
        "tags": [
          "replay"
        ]
      }
    },
...
    "/api/state_at": {
      "get": {
        "description": "The exact historical state at a dataset instant, for scrubbing.",
        "operationId": "state_at_api_state_at_get",
        "parameters": [
          {
            "in": "query",
            "name": "plant_id",
            "required": true,
            "schema": {
              "enum": [
                "ai4i",
                "ims"
              ],
              "title": "Plant Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "dataset_ts",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "Dataset Ts",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlantSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "summary": "State At",
        "tags": [
          "plants"
        ]
      }
    },
...
  }
}